[perl #105926] Make ‘require 6’ a syntactic special case
This commit moves the special-casing for ‘require 6’ from pp_require
(run time) to op.c (optree-building time).  If the argument to require
is a constant (not counting folded constants) that is a number or a
vstring, the require op is flagged as being a version check, with the
newly-added OPpREQUIRE_VER flag.

This means that ‘require $ver’ no longer works, which could be consid-
ered an incompatible change (some tests had to change), but the bug
fix below makes it worth it.

Since CORE::require($_[0]) will no longer work from an override (actu-
ally a callback; see perl #108286), overrides are no longer called for
version checks like ‘require 5’.  The tests for this are marked to-do
for now, as they will be re-enabled (with ‘use feature "overrides"’)
when the ‘overrides’ feature comes into effect.

I left the lexing part of require as it was, since, although it is
clearly buggy, it is not clear exactly how things are supposed to
work.  See perl #107004.

This is the bug fix we get as a result:

These two lines:

    $x = "6"; require $x;
    $x = "6"; 0+$x; require $x;

should behave exactly the same way, but the latter was a version
check.  Now they are both file names.  And this no longer happens:

    $ perl5.12.4 -le '$x = "34 cabbages"; 0+$x; require $x'
    Invalid version format (non-numeric data) at -e line 1.

In this case, "34 cabbages" should obviously be a file name, but its
numeric representation was causing it to be treated as a version num-
ber; yet the version-parsing code used the string representation.