Add a prototype for CORE::undef
Instead of calling ck_lfun on an undef op and marking it with S, we
can just mark it with R, which has the same effect.

ck_fun calls scalar(kid) on the child op if regen/opcodes indicates S
for the argument.  ck_lfun calls ck_fun and then modkids on its return
value, which in turn does op_lvalue(kid, ...).

If we use R, ck_fun does both scalar() and op_lvalue(), saving two
function calls, and a special case in core_prototype.

(There is still a special case in core_prototype, but a smaller one
than we would need otherwise.)