Make if hide else and elsif from PL_keyword_plugin
We do this by introducing two new states for PL_expect: XIFCOND and
XELSE.  This allows the lexer to rememeber by the time it reaches
the ‘}’ in ‘if(...){...}’ that it was parsing an ‘if’ statement.
PL_expect is XIFCOND between ‘if’ or ‘elsif’ and the following ‘(’.
The code for ‘(’ now pushes a PL_expect value on to PL_lex_brackstack,
just like ‘[’ and ‘{’.  ‘)’ now pops the brackstack.  So PL_expect
is also XIFCOND between the ‘)’ and ‘{’.  ‘{’ pushes XELSE on to the
brackstack if PL_expect was XIFCOND, so that XELSE will be popped off
the stack and put into PL_expect when ‘}’ is encountered.  ‘else’ sets
PL_expect to XBLOCK.

This means that any code looking for the beginning of a statement must
check for XELSE as well as XSTATE, so I put that in a macro.

The code for the keyword plugin is now skipped if PL_expect is XELSE
and the next token is ‘else’ or ‘elsif’.