[perl #112924] [PATCH] Clarify test instructions in INSTALL
While trying to track down a failed test, I found the instructions in
the INSTALL file to be less than optimal.

This patch re-orders the suggestions to put the ones most likely to work
first, clarifies which directory you should be in, and moves a "see-also"
type reference to where it is more likely to be useful.

The catalyst for this was that

    ./perl -MTestInit cpan/Archive-Extract/t/01_Archive-Extract.t

fails because it can't find strict.pm.  (See [perl #1122926].)  Since it's
probably not the only such case, I thought it useful and prudent to
advise using t/harness instead as a first recourse.