Make for(each) hide continue from PL_keyword_plugin
This uses the recently-added XLOOPCOND PL_expect state, which allows
the lexer to remember that it is parsing a while loop by the time it
reaches ‘}’, extending it to for(each).

Since for(;;) doesn’t take a while loop, a ‘;’ directly inside the
parentheses (i.e., not nested inside more brackets of any kind)
changes the PL_expect value stored on the brackstack.