handy.h: New defn of isOCTAL_A() to free up bit
The new definition is likely slightly faster, as it replaces an array
lookup with a mask.

Comments are also added, listing the other possible candidates for this
treatment, though the speed differential is unclear as they would also
add an extra test.

A U32 is used to store the information about the various properties for
a character.  This frees up one bit of that for future other use.