perlebcdic: Narrow table to fit in 79 columns
This uses the abbreviations for control characters available officially
in Unicode 6.1, instead of the much longer names, thus enabling this
table to be narrowed.  With some space changes, it will fit in 79
columns.  The differences look large, but most go away under 'diff -b'.

The recipes are changed to correspond.