mktables: Defer accepting Unicode's BELL character
As soon as 5.17 started, this created a failing test, which was to
remind us that the deprecation cycle was over, and that the name BELL
should be now used for U+1F514 instead of U+0007.  See commit messages
for 3ffed8c228f4a59b1527ae7ed584ca21386420f7 and
10914c783fe2ea3ee73a870599f30cedb7de96d0

This change puts off the failure unless we try to ship a 5.18 without
dealing with the issue.  I'm not quite ready to deal with it, but will
do so shortly, long before 5.18.