Revert "Fix mktables bug due to the previous overload fix"
mktables had unknowingly been relying on a bug in the overloading code.
That bug was fixed by commit f041cf0f9c6469c41de8b73d5f7b426710c3ff8b.
Commit 5f9f83be9cdcd54449f7f40db078fe367d780475 is a minimal commit to
get mktables to pass its tests as a result of the fixed bug.  However,
it did not address the underlying problem, which doesn't show up in the
typical tests, but does occur when tracing is added or things go wrong
and mktables tries to output any of a number of messages, which fail
because there is no ".=" operator.  A previous commit added the proper
overloaded ".=", and so this one is no longer needed, and would be
confusing to someone who doesn't know the history.