Fix up vms/vms.c for C++.
Almost all of this consists of casting the return values from
PerlMem_malloc, which inexpicably we weren't doing.

There are also some prototypes that have to be at file scope and
declare with extern "C".

And there was some embarrassing dead code in Perl_my_localtime
that the C++ compiler spotted.