Make configure.com ready for C++.
Building on Peter Prymmer's work from some years ago, this gets
the configuration process on VMS up to speed on the native C++
compiler.  Quite a bit of it involves tweaking the configuration
probes to work under C++, plus a variety of other nits.

TODO:  This should probably be made to work as -Dcc=cxx rather
than -Dusecxx as it does now, but it's still pretty experimental.