Revert "t/porting/checkcase.t should skip directories"
This reverts commit f86fc3ff06fb5e5bae2719b5601eb7a094fe014b.

David and I had discussed this commit, along with some others on IRC,
and we felt the test was wrong and the commit was good.  Unfortunately,
problems were reported only later, after Version-Requirements had
entered the core, so undoing the test would have required an
exception be made for lib/Version and lib/version conflicting.

Now that Version-Requirements has left the core, this test can be
restored.