In Porting/cmpVERSION.pl, increment $count in one place, unconditionally.
Previously it was only incremented if it was being used (as part of generating
TAP output). However, this was done as ++$count, necessitating using a printf
statement. By moving the increment to one place we can use print instead,
which is simpler, and avoid the risk of forgetting to increment $count if we
add a new block for different test output.