Teach Porting/cmpVERSION.pl about .pm files extracted from .pm_PL files.
Previously these were overlooked - the generated .pm file corresponding to
an .xs file would be found on disk, but as that file was not in git, the
original $VERSION could not be determined and it would be skipped. Now the
correct master file is checked, and failed if it is not updated.

Without this change, t/porting/cmp_version.t currently fails for a build
using -Dmksymlinks, due to ext/DynaLoader/dl_vms.xs having recent changes.
The test changes directory to the original git checkout, and if that
checkout is clean, there is *no* ext/DynaLoader.pm file there, which meant
that the previous version of pm_file_from_xs() would die because it could
not figure out which Perl file relates to dl_vms.xs