Do away with stashpv_hvname_match
For some reason this is listed in the API, even though it is not docu-
mented and is only available under ithreads.

It was added by commit ed221c5717, which doesn’t explain why it needed
to be part of the API.  (Presumably because a public macro used it,
even though there are better ways to solve that.)

It is unused on CPAN and (now) in core, so there is no reason
to keep it.