Fix bad assertions in pp_ctl.c:pp_caller
If a stash is freed or undefined by code that it calls, then pp_caller
can encounter a freed stash or nameless stash.  HvNAME_HEK returns
null for a freed scalar and for a hash without a name.  So CopSTASHPV
will also be null in that case, which is expected.  If the hash is
undefined, it may not be SvOOK any more, and if it is freed it defi-
nitely won’t be.

The following code *does* assume, however that CopSTASH is non-null,
so assert that.