Change B::COP::stashlen to stashoff
This was brought up in ticket #78742.

The stashlen method has never been in a stable release, and no longer
exists, as of d4d03940c, since it is dependent on a define that
d4d03940c removed.

So this commit removes stashlen from B.xs and adds stashoff in its
place, since this is what B::C needs.

It also adds a few basic tests for the stash and stashpv methods.